// packages/smartsale-content/scripts/datocms/graphql/generated/cms/schema.ts
export type Scalars = {
  Boolean: boolean
  BooleanType: any
  CustomData: any
  DateTime: any
  Float: number
  FloatType: any
  Int: number
  IntType: any
  ItemId: any
  JsonField: any
  MetaTagAttributes: any
  String: string
  UploadId: any
}
// ...
export interface ColorField {
  alpha: Scalars['IntType']
  blue: Scalars['IntType']
  cssRgb: Scalars['String']
  green: Scalars['IntType']
  hex: Scalars['String']
  red: Scalars['IntType']
  __typename: 'ColorField'
}
// ...
export interface FileField {
  _createdAt: Scalars['DateTime']
  /** The DatoCMS URL where you can edit this entity. To use this field, you need to set a X-Base-Editing-Url header in the request */
  _editingUrl: Scalars['String'] | null
  _updatedAt: Scalars['DateTime']
  alt: Scalars['String'] | null
  author: Scalars['String'] | null
  basename: Scalars['String']
  blurUpThumb: Scalars['String'] | null
  blurhash: Scalars['String'] | null
  colors: ColorField[]
  copyright: Scalars['String'] | null
  customData: Scalars['CustomData']
  exifInfo: Scalars['CustomData']
  filename: Scalars['String']
  focalPoint: focalPoint | null
  format: Scalars['String']
  height: Scalars['IntType'] | null
  id: Scalars['UploadId']
  md5: Scalars['String']
  mimeType: Scalars['String']
  notes: Scalars['String'] | null
  responsiveImage: ResponsiveImage | null
  size: Scalars['IntType']
  smartTags: Scalars['String'][]
  tags: Scalars['String'][]
  thumbhash: Scalars['String'] | null
  title: Scalars['String'] | null
  url: Scalars['String']
  video: UploadVideoField | null
  width: Scalars['IntType'] | null
  __typename: 'FileField'
}
// ...
export interface ResponsiveImage {
  alt: Scalars['String'] | null
  aspectRatio: Scalars['FloatType']
  base64: Scalars['String'] | null
  bgColor: Scalars['String'] | null
  height: Scalars['IntType']
  sizes: Scalars['String']
  src: Scalars['String']
  srcSet: Scalars['String']
  title: Scalars['String'] | null
  webpSrcSet: Scalars['String']
  width: Scalars['IntType']
  __typename: 'ResponsiveImage'
}

export interface SeoField {
  description: Scalars['String'] | null
  image: FileField | null
  noIndex: Scalars['BooleanType'] | null
  title: Scalars['String'] | null
  twitterCard: Scalars['String'] | null
  __typename: 'SeoField'
}
// ...
export interface UploadVideoField {
  alt: Scalars['String'] | null
  blurUpThumb: Scalars['String'] | null
  blurhash: Scalars['String'] | null
  duration: Scalars['Int'] | null
  framerate: Scalars['Int'] | null
  height: Scalars['IntType']
  mp4Url: Scalars['String'] | null
  muxAssetId: Scalars['String']
  muxPlaybackId: Scalars['String']
  streamingUrl: Scalars['String']
  thumbhash: Scalars['String'] | null
  thumbnailUrl: Scalars['String']
  title: Scalars['String'] | null
  width: Scalars['IntType']
  __typename: 'UploadVideoField'
}

export type VideoMp4Res = 'low' | 'medium' | 'high'

export interface focalPoint {
  x: Scalars['FloatType']
  y: Scalars['FloatType']
  __typename: 'focalPoint'
}